+++ CHANGELOG.md
@@ -5,9 +5,90 @@
 
 ### Fixes
 - Function guard should also allow `Nx.Tensor` when the corresponding input argument is `Evision.Mat.maybe_mat_in()`.
+- [Evision.Mat] `Evision.Mat.quicklook/1`  should also check the number of channels is one of `[1, 3, 4]` when `dims == 2`.
+- [c_src] `evision_cv_mat_broadcast_to` should call `enif_free((void *)dst_data);` if `void * tmp_data = (void *)enif_alloc(elem_size * count_new_elem);` failed.
 
 ### Changed
 - [Docs] Example Livebooks is now included in docs as extras.
+- [Evision.Mat] `Evision.Mat.roi/{2,3}` now supports Elixir Range.
+- [Evision.Mat] Implemented Access behaviour. For example,
+
+  ```elixir
+  iex> img = Evision.imread("test/qr_detector_test.png")
+  %Evision.Mat{
+    channels: 3,
+    dims: 2,
+    type: {:u, 8},
+    raw_type: 16,
+    shape: {300, 300, 3},
+    ref: #Reference<0.809884129.802291734.78316>
+  }
+
+  # Same behaviour as Nx. 
+  # Also, img[0] gives the same result as img[[0]]
+  # For this example, they are both equvilent of img[[0, :all, :all]]
+  iex> img[[0]]
+  %Evision.Mat{
+    channels: 3,
+    dims: 2,
+    type: {:u, 8},
+    raw_type: 16,
+    shape: {1, 300, 3},
+    ref: #Reference<0.809884129.802291731.77296>
+  }
+
+  # same as img[[0..100, 50..200, :all]]
+  # however, currently we only support ranges with step size 1
+  #
+  # **IMPORTANT NOTE**
+  #
+  # also, please note that we are using Elixir.Range here
+  # and Elixir.Range is **inclusive**, i.e, [start, end] 
+  # while cv::Range `{integer(), integer()}` is `[start, end)`
+  # the difference can be observed in the `shape` field
+  iex> img[[0..100, 50..200]]
+  %Evision.Mat{
+    channels: 3,
+    dims: 2,
+    type: {:u, 8},
+    raw_type: 16,
+    shape: {101, 151, 3},
+    ref: #Reference<0.809884129.802291731.77297>
+  }
+  iex> img[[{0, 100}, {50, 200}]]
+  %Evision.Mat{
+    channels: 3,
+    dims: 2,
+    type: {:u, 8},
+    raw_type: 16,
+    shape: {100, 150, 3},
+    ref: #Reference<0.809884129.802291731.77297>
+  }
+
+  # for this example, the result is the same as `Evision.extractChannel(img, 0)`
+  iex> img[[:all, :all, 0]]
+  %Evision.Mat{
+    channels: 1,
+    dims: 2,
+    type: {:u, 8},
+    raw_type: 0,
+    shape: {300, 300},
+    ref: #Reference<0.809884129.802291731.77298>
+  }
+  iex> img[[:all, :all, 0..1]]
+  %Evision.Mat{
+    channels: 2,
+    dims: 2,
+    type: {:u, 8},
+    raw_type: 8,
+    shape: {300, 300, 2},
+    ref: #Reference<0.809884129.802291731.77299>
+  }
+
+  # when index is out of bounds
+  iex> img[[:all, :all, 42]]
+  {:error, "index 42 is out of bounds for axis 2 with size 3"}
+  ```
 
 ## v0.1.11 (2022-10-13)
 [Browse the Repository](https://github.com/cocoa-xu/evision/tree/v0.1.11) | [Released Assets](https://github.com/cocoa-xu/evision/releases/tag/v0.1.11)